+++ components/questions/ToastUiCreate/style.js
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+export const BtnContainer = styled.div`
+  width: 21.875rem;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  margin-top: 1.5rem;
+`;
+
+export const Btn = styled.button`
+  width: 8.75rem;
+  height: 3rem;
+  background-color: #f5f5f7;
+  /* border: 3px solid #FFFFFF; */
+  /* border: none; */
+  box-sizing: border-box;
+  border-radius: 1.5625rem;
+  cursor: pointer;
+`;
