+++ components/questions/ToastUiUpdate/index.jsx
@@ -11,7 +11,7 @@ import { useRouter } from "next/router";
 import Prism from "prismjs";
 import React, { useCallback, useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import styled from "styled-components";
+import * from './style';
 
 import { API_URL } from "../../config";
 import { setDesc } from "../../modules/editor";
@@ -138,22 +138,3 @@ function ToastUiUpdate({ id, title, desc, tags, token }) {
 }
 
 export default React.memo(ToastUiUpdate);
-
-const BtnContainer = styled.div`
-  width: 21.875rem;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  margin-top: 1.5rem;
-`;
-
-const Btn = styled.button`
-  width: 8.75rem;
-  height: 3rem;
-  background-color: #f5f5f7;
-  /* border: 3px solid #FFFFFF; */
-  /* border: none; */
-  box-sizing: border-box;
-  border-radius: 1.5625rem;
-  cursor: pointer;
-`;
